Move `dendritic_layers` and `dendrite_segments` to new `modules` directory.
