+++ nupic/research/frameworks/dendrites/modules/__init__.py
@@ -1,6 +1,6 @@
 # ----------------------------------------------------------------------
 # Numenta Platform for Intelligent Computing (NuPIC)
-# Copyright (C) 2020, Numenta, Inc.  Unless you have an agreement
+# Copyright (C) 2021, Numenta, Inc.  Unless you have an agreement
 # with Numenta, Inc., for a separate license for this software code, the
 # following terms and conditions apply:
 #
@@ -19,20 +19,11 @@
 # http://numenta.org/licenses/
 # ----------------------------------------------------------------------
 
-from .modules import *
-from .plotting_utils import (
-    plot_dendrite_activations,
-    plot_dendrite_overlap_matrix,
-    plot_entropy_distribution,
-    plot_mean_selected_activations,
-    plot_overlap_scores_distribution,
-    plot_percent_active_dendrites,
-)
-from .utils import (
-    dendrite_duty_cycle,
-    dendrite_overlap,
-    dendrite_overlap_matrix,
-    entropy,
-    mean_selected_activations,
-    percent_active_dendrites,
+from .dendrite_segments import DendriteSegments
+from .dendritic_layers import (
+    AbsoluteMaxGatingDendriticLayer,
+    AbsoluteMaxGatingDendriticLayer2d,
+    BiasingDendriticLayer,
+    GatingDendriticLayer,
+    GatingDendriticLayer2d,
 )
