+++ nupic/research/frameworks/dendrites/modules/dendrite_segments.py
