+++ nupic/research/frameworks/dendrites/modules/dendritic_layers.py
